fix(aggregate): cast INT64_MIN/MAX to double to avoid implicit conversion warnings

+++ src/aggregate_function.cpp
@@ -312,8 +312,11 @@ void CustomAggregate::JSValueToSqliteResult(sqlite3_context *ctx,
     }
   } else if (value.IsNumber()) {
     double num_val = value.As<Napi::Number>().DoubleValue();
-    if (floor(num_val) == num_val && num_val >= INT64_MIN &&
-        num_val <= INT64_MAX) {
+    // Note: We cast INT64_MIN/MAX to double to avoid implicit conversion
+    // warnings
+    if (floor(num_val) == num_val &&
+        num_val >= static_cast<double>(INT64_MIN) &&
+        num_val <= static_cast<double>(INT64_MAX)) {
       sqlite3_result_int64(ctx, static_cast<sqlite3_int64>(num_val));
     } else {
       sqlite3_result_double(ctx, num_val);
